Fix corrupted _FillValue metadata on pop/rain/t2m (kind mismatch)
FillValue, FillValue_rain, and FillValue_temp were declared as
single-precision real, but get passed to nf90_def_var_fill for pop/
rain/t2m -- all declared xtype = nf90_double in the output file. The
netCDF C API interprets the fill_value pointer according to the
variable's own declared type, so it read 8 bytes starting at a 4-byte
Fortran variable's address: 4 real bytes plus 4 bytes of whatever
happened to be adjacent in memory. Declared _FillValue ended up as
nonsense (e.g. pop: 1.66e-314, rain: -5.13e+37, t2m: -5.59e+29)
instead of the real input sentinels (-99999, -9999, -9999).

Found while extending the same fix already applied to Age's
_FillValue (previous commit) to rain/t2m -- turned out to be the same
root cause, and pop was affected too.

Widened all three to double precision, matching the double xtype of
every field they're attached to (consistent with how Age's fix used
nf90_fill_double directly rather than a mismatched-kind variable).
Reading the real input file's _FillValue via nf90_get_att into these
variables, and the where(...) mask-setting comparisons against
pop_dens/rainfall/t2m, were never affected by this bug (ordinary
scalar assignment and comparison don't require matching kinds) --
only the raw-pointer nf90_def_var_fill call was.

Verified: full-length Kenya benchmark (1M agents, 1096 days) against
a worktree of the previous commit. pop shows zero cdo diffn
differences (the model already zeroes masked-out pop cells before
writing, so no fill-value cells exist in its output either way).
rain/t2m show diffn differences on every timestep, but confirmed via
direct ncks spot-check this is the same masking artifact as Age's fix:
raw underlying bytes are identical between builds at masked cells
(e.g. -9999 at a known ocean/edge grid point) -- the reference build's
garbage _FillValue just doesn't match, so it shows -9999 as literal
data instead of masking it, while the fixed build correctly excludes
it. EIR, HA, Nagent, V, imm, wperm, wpond, Age (already fixed) all
show zero differences.

+++ src/mo_const.f90
@@ -35,10 +35,14 @@ MODULE mo_const
     integer :: ImmVarID, ncImmID   ! IDs for immunity forcing file
 
 
-    real    :: eps = 1e-15  ! Numerical tolerance 
-    real    :: FillValue    ! Population density FillValue
-    real    :: FillValue_rain ! Rainfall Fill Value
-    real    :: FillValue_temp ! Temperature Fill Value
+    real    :: eps = 1e-15  ! Numerical tolerance
+    ! double precision (not real): these get passed to nf90_def_var_fill for
+    ! nf90_double output variables (pop/rain/t2m), which reads 8 bytes at the
+    ! given address -- a 4-byte real here would leave the library reading 4
+    ! bytes of adjacent memory as the other half of the value, corrupting it.
+    double precision :: FillValue    ! Population density FillValue
+    double precision :: FillValue_rain ! Rainfall Fill Value
+    double precision :: FillValue_temp ! Temperature Fill Value
     real, parameter :: pi = 4*atan(1.) ! Pi 
     real, parameter :: Re = 6371.009   ! Mean radius of the Earth [km]
